Add `plan` to optional properties
We want to pass `plan: { branch: "some name" }` as part of the payload,
but since this isn't a whitelisted property, AmplitudeAPI ends up
creating accessors dynamically. This has a bug: only first
initialization actually adds `plan` to the payload, subsequent events
don't have this value. To workaround it, we're extending the list of
properties itself, which isn't possible to do with current config setup.